+++ scripts/build-pkg.sh
@@ -31,6 +31,8 @@ readme_path = os.path.join(root, "README.md")
 
 files = [
     ("/usr/local/bin/gentlsa", bin_path, 0o755),
+    ("/usr/local/share/man/man1/gentlsa.1", os.path.join(root, "man/gentlsa.1"), 0o644),
+    ("/usr/local/share/man/man5/gentlsa.5", os.path.join(root, "man/gentlsa.5"), 0o644),
     ("/usr/local/share/licenses/gentlsa/LICENSE", license_path, 0o644),
     ("/usr/local/share/doc/gentlsa/README.md", readme_path, 0o644),
 ]